🚚 rename: import 수정
